Implemented SQLite for persistent storage.
This removed a Cogs' save and load related methods, instead giving a subclass-able class that essentially act as a row in a table. This should provide a sufficient abstraction for my purposes.

The bot key, owner(s) list, and other one off data will still live in a json file though.

Also modified logging so that the general log / console won't be filled with library logging actions. Moved the general log into a file with a set name, while warnings and up get a dated file.