Refactored main.py into system.py and discord_bot.py.
Most of the contents of main.py got moved to system.py. Notable exclusions include bits relating to command prefixes and disabled commands, as both these functionalities have been made obsolete application commands. System is also currently no longer a cog and is loaded with the bot.
The "main" function got moved to discord_bot.py, which is now acts as the runner. It currently has two command line arguments to control logging, but it may be expanded in the future.

Error handling has been removed from cog_manager.py to system.py

make_error now returns an ErrorEmbed which is just a standard discord Embed but with a boolean value of false so functions can communicate failures through the embed type.
Added a function for use as a decorator that replaces a function's annotations with a discord Option with the description pulled from the docstring. This is so arguments in the slash commands in discord show helpful information.