Implemented encrypting some Storable attributes.
Attributes named in Storable.encrypt_attrs will be encrypted for saving in the database. The key needs to be provided from an outside source using Storable.set_key(key). The primary key may not be encrypted as it is used for the nonce.

Also added a modal in system to unlock Storable classes.