Refactored _with_notif and added a notification to code sharer.
Moved hoyolab._._with_notif to system.do_and_dm and system.send_dm. It was already used multiple times, and it would be likely to be duplicated again in the future in other contexts.

Added sending code redemption notification for the user who shares a code (if their notif_codes is True). It wasn't clear to the user if the code was redeemed for them as well.