Caught an exception that only appears on Pi
The genshin.client().get_partial_genshin_user() works fine when running on my computer, but raises a (generic) Exception when running on my Raspberry Pi. Since _check is primarily for validating the cookies and the exception seems to only occur with valid cookies, the function shouldn't break when it occurs and should still return a "passing" value.