Added customizable log get with `/vclog get` and added more params to the other vclog commands.
`/vclog get` has a boolean parameter for each type of voice state change that turns that on or off (default).
Added a `time_format` parameter to all the vclog commands that lets the caller set which time format to use. `/vclog all` and `/vclog get` also got the `remove_dupes` and `remove_undo` parameters to pass on to `_vc_log_embed`.
Also updated / expanded the docstrings and some type hints were simplified with the `VOICE_STATE_CHANNELS` constant.