Made Storable ensure table exists before saving.
In a scenario where the table doesn't exist and multiple objects tries to save, the table could be told to be created multiple times, which could have unintended consequences.
This basically adds a lock to Storable's save.

Adds possible issue if the table is deleted after the first save. However, this isn't an option in the wrapper and breaks assumptions.